Reimplement TSourceFileInfo.FileTypeInfo[] property
Changed the field and read/write accessor for this property to store the
property values in a dictionary instead of in a fixed size array.
Functionality was changed only in as much that attempts to access an
"array" value will now fail with an exception if an attempt is made to
get a value that has not been previously set. Before garbage results
would be returned.

This change was made to enable fewer file types than the maximum to be
supported in the filter string created by TSourceFileInfo.FilterString.