Add new TSourceFileInfo.FileTypeFromFilterIdx method
New method to get the file type associated with a given index within a
filter string.

To implement this many changes were made to the internals of
TSourceFileInfo, the main one of which was that the filter string is
now generated every time the FileTypeInfo property is updated instead of
the filter string being built on request.