Reimplement TSaveSourceMgr.FileTypeFromFilterIdx
This method was changed to simply call
TSourceFileInfo.FileTypeFromFilterIdx for the currently selected filter
in the associated dialogue's filter string, instead of calculating the
value locally.

Note that the new method is much more resilient to future changes than
the original implementation which made assumptions about a one to one
relationship between filter indexes and file types.