Rewrite of USaveInfoMgr ready for extra file types
The unit was rewritten to use the TSaveSourceDlg dialogue box instead of
the simple TSaveDialogEx.

Nearly all code was rewritten, although, despite using a different
dialogue box, only RTF files are supported for output. The only
difference being that highlighting the source code in the output can now
be switched off.

The code was rewritten to make it easier to add support for other file
formats.