Add support to TCSS for margin & padding units
Modified all TCSS.PaddingProp and TCSS.MarginProp overloaded methods to
add a new optional parameter to specify which length units to use.
Previously only px was permitted.

In order to permit fractional lengths to be specified for use with
relative units, length parameters were changed to floating point from
Integer. Non-Integer values are truncated to a maximum of 2 decimal
places.