Add new Markdown source file type.
Added sfMarkdown element to TSourceFileType enumerated type.

The new sfMarkdown element broke TSaveSnippetMgr.GetFileTypeDesc and
TSaveUnitMgr.GetFileTypeDesc which assumed that all TSourceFileType
values were supported, and so don't check for any file type that may not
be supported. To enable such checks to be made a new
TSourceFileInfo.SupportsFileType method was added to check if a file
type is supported.

Updated the Code Formatting tab of the Preferences dialogue box with the
options to set Markdown as the default source file type.