Fix snippet info preview bug for ANSI encodings
When either the plain text or Markdown file types were selected in the
Save Snippet Information dialogue box and the ANSI encoding was also
selected, the snippet displayed when the output ws previewed could
differ from that written to file. This occured when the snippet
contained characters that couldn't be rendered correctly in ANSI: the
preview would show the correct snippet (rendered in Unicode) while the
snippet with incorrectly translated characters was written to file.

This bug was fixed so that the snippet previewed was also in ANSI
encoding, meaning that any encoding errors show up there exactly as they
will be written to file.

Fixes #164