Standardise and improve file encoding names
File encoding names are now determined by the encoding type and are no
longer specified by the code that displays TSaveSourceDlg dialogue
boxes. This not only standardises the encoding names but removes quite a
bit of duplication and simplifies code that saves snippets, snippet
information and units.

The system default ANSI encoding was previously displayed as "ANSI
(Default)". This has been changed to "ANSI Code Page 999", where 999 is
the default system code page for the user's locale.