Change default filename in Save Snippet Info dlg
The default file name was always "SnippetInfo", which could lead to
accidental overwrites of previously save information. The default was
changed to be a name based on the snippet display name.

Modified TSourceFileInfo with new property to prevent default file names
from being converted as necessary to make them safe Pascal identifiers:
we don't want this to happen in the Save Snippet Information dialogue.