Remove preferences dependency from USourceGen unit
Modified TSourceGen.UnitAsString to get information aboute compiler
warnings via a new parameters instead of from the Preferences object.
This was the only dependency on the Preferences object in the unit.

Calling code was modified to pass the required warnings to
TSourceGen.UnitAsString as a parameter. This calling code now gets the
value from the preferences object.

Fixes #167